+++ example/json/json.ts
@@ -3,7 +3,7 @@ import { Lexer, ELR } from "../../src";
 const lexer = new Lexer.Builder()
   .ignore(Lexer.whitespaces) // ignore blank characters
   .define({
-    string: Lexer.stringLiteral({ double: true }), // double quote string literal
+    string: Lexer.stringLiteral(`"`), // double quote string literal
     number: /^-?(?:0|[1-9]\d*)(?:\.\d+)?(?:[eE][+-]?\d+)?/,
   })
   .define(Lexer.wordType("true", "false", "null")) // type's name is the literal value