+++ example/string-literal/string-literal.ts
@@ -3,12 +3,9 @@ import { Lexer } from "../../src";
 export const lexer = new Lexer.Builder()
   .ignore(Lexer.whitespaces)
   .define({
-    simpleStr: Lexer.stringLiteral({
-      single: true,
-      double: true,
-    }),
-    multilineStr: Lexer.stringLiteral({ back: true, multiline: true }),
-    customQuote: Lexer.stringLiteral({ quote: "*" }),
-    custom: Lexer.stringLiteral({ from: "^", to: "$$" }),
+    simpleStr: Lexer.stringLiteral(`'`).or(Lexer.stringLiteral(`"`)),
+    multilineStr: Lexer.stringLiteral("`", { multiline: true }),
+    customQuote: Lexer.stringLiteral(`*`),
+    custom: Lexer.stringLiteral(`^`, { close: "$$" }),
   })
   .build();