+++ CHANGELOG.md
@@ -1,11 +1,11 @@
 # CHANGELOG
 
-## v0.9.1
+## v0.10.0
 
 - Lexer
   - Feat: Add `Action.or` to cascade actions.
   - Add utils: `Lexer.whitespaces/comment`. #5
-  - Deprecate `Lexer.from_to`, use `Lexer.fromTo` instead.
+  - **Breaking Change**: new `Lexer.stringLiteral`.
 
 ## v0.9.0